feat: Move language select to header
